Quick fix to add homoscedasticity to torch functions (#85)
Co-authored-by: Alec Dunton <dunton1@llnl.gov>

> > > I didn't catch this gap when I wrote PR #80. Is `MuyGPyS.torch` not covered by any test cases? If not, we should add some to ensure that this sort of thing doesn't happen again.
> > 
> > 
> > We should probably put tests for `MuyGPyS.torch` in a `tests/torch.py` to start.
> 
> Let me confirm this - I am pretty sure I covered that module in the test bed.

There is a test in  `tests/api/torch.py`, but those tests do not get run by the CI because they require data that we can't host on github. What I'm talking about is a mechanical test called `tests/torch.py` that just checks that the `MuyGPyS.torch` stuff does what we expect (shapes, any output guarantees, optimization runs without breaking, etc). We'll need to add that to the `.github/workflows/develop-test.yml` CI as well so that it gets run automatically.